# Fix performance issue