--- a/src/DependencyEmitter/UsesDependencyEmitter.php
+++ b/src/DependencyEmitter/UsesDependencyEmitter.php
@@ -8,6 +8,9 @@ use Qossmic\Deptrac\AstRunner\AstMap;
 use Qossmic\Deptrac\AstRunner\AstMap\AstDependency;
 use Qossmic\Deptrac\Dependency\Dependency;
 use Qossmic\Deptrac\Dependency\Result;
+use function array_map;
+use function array_merge;
+use function explode;
 
 class UsesDependencyEmitter implements DependencyEmitterInterface
 {
@@ -18,12 +21,26 @@ class UsesDependencyEmitter implements DependencyEmitterInterface
 
     public function applyDependencies(AstMap $astMap, Result $dependencyResult): void
     {
+        $references = array_merge($astMap->getAstClassReferences(), $astMap->getAstFunctionReferences());
+        $referencesFQDN = array_map(
+            static function ($ref) {
+                return $ref->getTokenName()->toString();
+            },
+            $references
+        );
+
+        $FQDNIndex = new FQDNIndexNode();
+        foreach ($referencesFQDN as $reference) {
+            $path = explode('\\', $reference);
+            $FQDNIndex->setNestedNode($path);
+        }
+
         foreach ($astMap->getAstFileReferences() as $fileReference) {
             $dependencies = $fileReference->getDependencies();
             foreach ($fileReference->getAstClassReferences() as $astClassReference) {
                 foreach ($dependencies as $emittedDependency) {
                     if (AstDependency::USE === $emittedDependency->getType() &&
-                        $this->isFQDN($emittedDependency, $astMap)
+                        $this->isFQDN($emittedDependency, $FQDNIndex)
                     ) {
                         $dependencyResult->addDependency(
                             new Dependency(
@@ -38,28 +55,15 @@ class UsesDependencyEmitter implements DependencyEmitterInterface
         }
     }
 
-    protected function isFQDN(AstDependency $dependency, AstMap $astMap): bool
+    protected function isFQDN(AstDependency $dependency, FQDNIndexNode $FQDNIndex): bool
     {
         $dependencyFQDN = $dependency->getTokenName()->toString();
-        $references = array_merge($astMap->getAstClassReferences(), $astMap->getAstFunctionReferences());
-
-        $isFQDN = false;
-        $isNamespace = false;
-
-        foreach ($references as $reference) {
-            $referenceFQDN = $reference->getTokenName()->toString();
-
-            if ($referenceFQDN === $dependencyFQDN) {
-                $isFQDN = true;
-            }
-
-            if (0 === strpos($referenceFQDN, $dependencyFQDN, 0) &&
-                $referenceFQDN !== $dependencyFQDN
-            ) {
-                $isNamespace = true;
-            }
+        $path = explode('\\', $dependencyFQDN);
+        $value = $FQDNIndex->getNestedNode($path);
+        if (null === $value) {
+            return true;
         }
 
-        return !$isNamespace || $isFQDN;
+        return $value->isFQDN();
     }
 }