--- a/tests/DependencyEmitter/Fixtures/FQDNIsSubstring/FQDN/SomeClass.php
+++ b/tests/DependencyEmitter/Fixtures/FQDNIsSubstring/FQDN/SomeClass.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace FQDNIsSubstring\FQDN;
+
+class SomeClass {}