--- a/tests/DependencyEmitter/Fixtures/FQDNIsSubstring/Foo.php
+++ b/tests/DependencyEmitter/Fixtures/FQDNIsSubstring/Foo.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace FQDNIsSubstring;
+
+use FQDNIsSubstring\FQDN\Some;
+use FQDNIsSubstring\FQDN\SomeClass;
+
+final class Foo {
+    public function foo() {
+        new Some();
+        new SomeClass();
+    }
+}