Closes  #798

```
real    0m8,316s
user    0m7,007s
sys     0m1,129s
```

versus vanila 0.19.1^
```
real    7m38,165s
user    7m32,888s
sys     0m1,887s
```